Enemy Health Bar and Shell faction
Implemented the enemy health bar as well as a system to check the Faction/Tag of the tank that fired. This is made so that the enemy health bar only appears when it's the player that damaged the enemy.

Messy implementation to be fair, will desperately need refactoring. Works for now though.