Created first mission and performance upgrades
The basis for the first mission has been set, now all that's needed is to flesh it out, improve the performance on it by remaking the treeline, improving the scenery, and creating the mission scripts. Also needed to create the cinematic cutscene with timeline, maybe.

There are also some very nice performance upgrades which make the game not want to die the moment there are two AI moving at the same time.

Also the high amount of new files are due to the new imported asset.