First mission created
First mission created, performance is in the gutter but fuck that, I don't give a shiiieeet right now.

In the future, maybe multithread some of the calculations, in the present however, I don't have neither the will nor the time to do this shit bro, I just can't. It also implies refactoring quite a large chunk of how the AI works probably, and that's already delicate enough.

Also, maybe check out a way to increase view distance since I spent too much on that scenery.

And finally, add a skybox and out of bounds, as well as all of the fucking mission stuff code shit fuuuuuuuuuuuuck