First mission (#17)
First mission implemented with some performance problems. These will need to be ironed off, not to mention the entire game being re-done because of course I could just do that too, and it would be much easier in a way. Still, not bad for less than three months with a language I didn't know how to use, a game engine I had never used before, and one dimension more than what I was accustomed to. Farewell Steel Protocol, I'll come back sooner than you'll expect to finish your beauty, and to pay homage to my favorite game of all time, Seek And Destroy.

So this is it for the Alpha. The Alpha is dead, long live the Beta! I'm going to do some War Thunder bamboozling and instead of going to v1.0.0 I'm just going to keep the count and move to version 0.10.0, except the difference is that now on the release tags it's going to say beta instead of alpha.